you can die by bullet

+++ space-shooter/src/game.rs
@@ -47,7 +47,7 @@ pub struct Sprite {
     pub shape: Shape,
     pub sprite_type: SpriteType,
     pub controller: Option<Controller>,
-    pub bullet_interval_counter: i32,
+    pub bullet_interval_counter: usize,
 }
 
 #[derive(Clone)] // Only clone
